+++ tools/firefox.sh
@@ -3,17 +3,23 @@
 ## Variables definitions
 
 # Tool name
-TOOL_NAME=asdf
+TOOL_NAME=firefox
 
 # Tool extra information
 EXTRA_INFO=
 
 # Install methods by OS
-OS_METHODS="darwin:brew linux:brew"
+OS_METHODS="darwin:cask linux:apt"
 
 ## Pre-installation required steps
 preInstall() {
+
     info "Pre Install for $1"
+
+    if [[ $OSTYPE == linux* ]] ; then
+        sudo add-apt-repository ppa:mozillateam/firefox-stable -y
+        sudo apt update
+    fi
 }
 
 ## Post-installation required steps