+++ tools/sublime-text.sh
@@ -3,17 +3,24 @@
 ## Variables definitions
 
 # Tool name
-TOOL_NAME=asdf
+TOOL_NAME=sublime-text
 
 # Tool extra information
 EXTRA_INFO=
 
 # Install methods by OS
-OS_METHODS="darwin:brew linux:brew"
+OS_METHODS="darwin:cask linux:apt"
 
 ## Pre-installation required steps
 preInstall() {
+
     info "Pre Install for $1"
+
+    if [[ $OSTYPE == linux* ]] ; then
+        wget -qO - https://download.sublimetext.com/sublimehq-pub.gpg | sudo apt-key add -
+        echo "deb https://download.sublimetext.com/ apt/stable/" | sudo tee /etc/apt/sources.list.d/sublime-text.list
+        sudo apt-get update
+    fi
 }
 
 ## Post-installation required steps